Change interger to positive small integer

+++ backend/track_my_moves/models/account.py
@@ -5,6 +5,6 @@ class Account(models.Model):
     first_name = models.CharField(max_length=50)
     last_name = models.CharField(max_length=50)
     birthdate = models.DateField(auto_now=False, auto_now_add=False)
-    height = models.IntegerField()
-    weight = models.IntegerField()
+    height = models.PositiveSmallIntegerField()
+    weight = models.PositiveSmallIntegerField()
     country = models.CharField(max_length=50)
